rename "step" to steps update recommendation_steps.py
