--- features/step/recommendation_steps.py
@@ -1,55 +1,0 @@
-"""
-Recommendation Steps
-Steps file for recommendations.feature
-"""
-from os import getenv
-import json
-import requests
-from behave import *
-from compare import expect, ensure
-# from selenium.webdriver.common.by import By
-# from selenium.webdriver.support.ui import WebDriverWait
-# from selenium.webdriver.support import expected_conditions
-
-# WAIT_SECONDS = 30
-BASE_URL = getenv('BASE_URL', 'http://localhost:5000/')
-
-@given('the following recommendations')
-def step_impl(context):
-    """ Delete all Recommnedations and load new ones """
-    headers = {'Content-Type': 'application/json'}
-    context.resp = requests.delete(context.base_url + '/recommendations/reset', headers=headers)
-    expect(context.resp.status_code).to_equal(204)
-    create_url = context.base_url + '/recommendations'
-    for row in context.table:
-        data = {
-            "productid": row['productid'],
-            "suggestionid": row['suggestionid'],
-            "categoryid": row['categoryid'],
-            }
-        payload = json.dumps(data)
-        context.resp = requests.post(create_url, data=payload, headers=headers)
-        expect(context.resp.status_code).to_equal(201)
-
-@when('I visit the "home page"')
-def step_impl(context):
-    """ Make a call to the base URL """
-    context.driver.get(context.base_url)
-    #context.driver.save_screenshot('home_page.png')
-
-@then('I should see "{message}" in the title')
-def step_impl(context, message):
-    """ Check the document title for a message """
-    expect(context.driver.title).to_contain(message)
-
-@then('I should not see "{message}"')
-def step_impl(context, message):
-    error_msg = "I should not see '%s' in '%s'" % (message, context.resp.text)
-    ensure(message in context.resp.text, False, error_msg)
-
-@when('I set the "{element_name}" to "{text_string}"')
-def step_impl(context, element_name, text_string):
-    element_id = 'pet_' + element_name.lower()
-    element = context.driver.find_element_by_id(element_id)
-    element.clear()
-    element.send_keys(text_string)
